+++ taskiq_pg/__init__.py
@@ -1,4 +1,5 @@
 from taskiq_pg.broker import AsyncpgBroker
+from taskiq_pg.middlewares import OrderedRetryMiddleware
 from taskiq_pg.result_backend import AsyncpgResultBackend
 
 __version__ = "0.2.0"
@@ -6,4 +7,5 @@ __version__ = "0.2.0"
 __all__ = [
     "AsyncpgBroker",
     "AsyncpgResultBackend",
+    "OrderedRetryMiddleware",
 ]